% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gen_model}
\alias{gen_model}
\title{Generate malariasimple deterministic model}
\usage{
gen_model()
}
\description{
Generate malariasimple deterministic model
}
\examples{
gen <- gen_model()
params <- get_parameters() |> set_equilibrium(init_EIR = 5)
out <- run_simulation(gen,params)

}
