% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_daily_decay}
\alias{get_daily_decay}
\title{Daily Effective Decay}
\usage{
get_daily_decay(usage_mat, decay_mat)
}
\arguments{
\item{usage_mat}{Matrix of usage values for each distribution event and simulation day.}

\item{decay_mat}{Matrix of decay values for each distribution event and simulation day.}
}
\description{
Estimates the mean average intervention decay across all distribution events
}
