% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_decay_mat}
\alias{get_decay_mat}
\title{Decay matrix for ITN or SMC}
\usage{
get_decay_mat(
  days,
  n_days,
  gamman_itn = NULL,
  scale_smc = NULL,
  shape_smc = NULL,
  intervention = NULL
)
}
\arguments{
\item{days}{Days on which distribution events occur}

\item{n_days}{Length of simulation (days)}

\item{gamman_itn}{Half-life of ITN insecticide (days)}

\item{scale_smc}{Scale parameter of Weibull distribution defining decay of prophylactic protection of SMC}

\item{shape_smc}{Shape parameter of Weibull distribution defining decay of prophylactic protection of SMC}

\item{intervention}{"ITN" or "SMC"}
}
\description{
Produces an ij matrix representing the decay in efficacy of insecticide-treated nets (ITN) or
prophylactic protection (SMC) on day j for individuals in receipt of distribution event i.
}
