% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_spraying.R
\name{get_irs_usage_mat}
\alias{get_irs_usage_mat}
\title{Produces matrix of IRS usage}
\usage{
get_irs_usage_mat(days, coverages, n_days, distribution_type)
}
\arguments{
\item{days}{Days on which IRS is implemented}

\item{coverages}{Population coverage for each day}

\item{n_days}{Length of simulation in days}

\item{distribution_type}{"random" or "cohort" distribution of IRS}
}
\description{
Produces an ij matrix defining the proportion of the population using an IRS from distribution i on day j. Final column represents population with no IRS.
}
