% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bednets.R
\name{get_itn_usage_mat}
\alias{get_itn_usage_mat}
\title{Produce matrix of population ITN usage}
\usage{
get_itn_usage_mat(days, coverages, retention, n_days)
}
\arguments{
\item{days}{Vector of days on which ITN distribution events occur (required when continuous_distribution = FALSE)}

\item{coverages}{Vector detailing the proportion of the population receiving an ITN during each intervention (required when continuous_distribution = FALSE)}

\item{retention}{Average number of days a net is kept for}

\item{n_days}{Length of simulation}
}
\description{
Produces an ij matrix defining the proportion of the population using an ITN from distribution i on day j. Final column represents population with no ITN.
}
