% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{get_output_colnames}
\alias{get_output_colnames}
\title{Get column names for simulation output}
\usage{
get_output_colnames(sys, params)
}
\arguments{
\item{sys}{Dust system. Output from dust2::dust_system_create()}

\item{params}{Simulation parameters}
}
\description{
Get column names for simulation output
}
