% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_2d}
\alias{make_2d}
\title{Converts 3D to long 2D}
\usage{
make_2d(array_output)
}
\arguments{
\item{array_output}{Any 3D array}
}
\description{
Takes a 3D array simulation output with multiple particles, and melts the third dimension to produce a 2D dataframe.
}
\examples{
params <- get_parameters(stochastic = TRUE) |>
            set_equilibrium(init_EIR = 5)
long_output <- run_simulation(params, n_particles = 3) |> make_2d()
}
