% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malariasimple-package.R
\docType{package}
\name{malariasimple-package}
\alias{malariasimple}
\alias{malariasimple-package}
\title{malariasimple: A time-discrete compartmental model for malaria}
\description{
A faster, simpler implementation of malariasimulation built using the package 'odin.dust'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrc-ide/malariasimple}
  \item Report bugs at \url{https://github.com/mrc-ide/malariasimple/issues}
}

}
\author{
\strong{Maintainer}: Debbie Shackleton \email{dshackle@imperial.ac.uk}

Authors:
\itemize{
  \item Peter Winskill \email{p.winskill@imperial.ac.uk}
}

}
\keyword{internal}
