% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exports.R
\name{malariasimple_deterministic}
\alias{malariasimple_deterministic}
\title{Deterministic malaria model generator}
\value{
A `dust_system_generator` you can pass to `dust2::dust_system_create()`
}
\description{
Returns the odin‐compiled dust_system_generator for the deterministic model.
}
