% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run deterministic malariasimple}
\usage{
run_simulation(params, n_particles = 1, full_output = FALSE)
}
\arguments{
\item{params}{List of model parameters generated using get_parameters() and other helper functions}

\item{n_particles}{Number of simulations to perform}

\item{full_output}{Boolean variable stating whether model should output model outputs, or a selected few.}
}
\description{
Generates and runs the deterministic malariasimple model
}
\examples{
params <- get_parameters() |>
          set_equilibrium(init_EIR = 10)
simulation_output <- run_simulation(params)
}
