% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_bednets.R
\name{set_bednets}
\alias{set_bednets}
\title{Set ITN parameters}
\usage{
set_bednets(
  params,
  continuous_distribution = FALSE,
  daily_continuous_cov = NULL,
  days = NULL,
  coverages = NULL,
  gamman = 2.64 * 365,
  retention = 3 * 365,
  dn0 = 0.41,
  rn = 0.56,
  rnm = 0.24
)
}
\arguments{
\item{params}{malariasimple parameters}

\item{continuous_distribution}{Is ITN distribution continuous? If FALSE, distribution is assumed to occur in discrete events.}

\item{daily_continuous_cov}{Vector of daily ITN coverage (required when continuous_distribution = TRUE). A single value is also accepted}

\item{days}{Vector of days on which ITN distribution events occur (required when continuous_distribution = FALSE)}

\item{coverages}{Vector detailing the proportion of the population receiving an ITN during each intervention (required when continuous_distribution = FALSE)}

\item{gamman}{Half-life of ITN insecticide (days)}

\item{retention}{Average number of days a net is kept for}

\item{dn0}{Probability of mosquito dying upon an encounter with ITN (max)}

\item{rn}{Probability of repeating behaviour with ITN (max)}

\item{rnm}{Probability of repeating behaviour with ITN (min)}
}
\value{
Updates the input parameter list to include ITN parameters
}
\description{
Set ITN parameters
}
\examples{
n_days <- 500
#Discrete distribution scenario
discrete_itn_params <- get_parameters(n_days = n_days) |>
  set_bednets(days = c(50,100,200),
          coverages = c(0.2,0.5,0.1)) |>
  set_equilibrium(init_EIR = 10)

#Continuous distribution scenario
coverageserage <- 0.2 + 0.15*(sin(2 * pi * (1:n_days / 365)) + 1)
discrete_itn_params <- get_parameters(n_days = n_days) |>
  set_bednets(continuous_distribution = TRUE,
          daily_continuous_cov = coverageserage) |>
  set_equilibrium(init_EIR = 10)
}
