% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_smc.R
\name{set_drug_params}
\alias{set_drug_params}
\title{Set drug parameter defaults}
\usage{
set_drug_params(drug)
}
\arguments{
\item{drug}{A character string specifying the drug type. One of:
\itemize{
  \item "SP_AQ": Sulfadoxine-pyrimethamine with Amodiaquine [0.9, 0.32, 4.3, 38.1]
  \item "AL": Artemether-Lumefantrine [0.95,0. 05094, 11.3, 10.6]
  \item "DHA_PQP": Dihydroartemisinin-Piperaquine [0.95, 0.09434, 4.4, 28.1]
}}
}
\value{
A vector of drug parameters for the specified drug in the format
[drug_efficacy, drug_rel_c, prophylaxis_shape, prophylaxis_scale]
}
\description{
Sets default parameters for commonly used SMC drugs. Values from SI of Commun. 5:5606 doi: 10.1038/ncomms6606 (2014).
}
\examples{
set_drug_params(drug = "SP_AQ")
}
