% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_equilibrium.R
\name{set_equilibrium}
\alias{set_equilibrium}
\title{Set equilibrium}
\usage{
set_equilibrium(params, init_EIR)
}
\arguments{
\item{params}{List of parameters}

\item{init_EIR}{Value of EIR at equilbrium}
}
\description{
Updates parameter list to include equilibrium values for a given EIR (entomological innoculation rate) for the malariasimple parameter list.
This function also includes some 'finishing touches' for the smooth running of the simulation and so is essential that this function is performed last.
}
\examples{
params <- get_parameters() |>
            set_equilibrium(init_EIR = 6)
}
