% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{set_rainfall_manual}
\alias{set_rainfall_manual}
\title{Add manual rainfall forcing}
\usage{
set_rainfall_manual(params, cc_ts)
}
\arguments{
\item{params}{Other malariasimple parameters}

\item{cc_ts}{Carrying capacity time series. Vector of daily carrying capacity. Length must equal or exceed params$n_days}
}
\value{
Updates the input parameter list to include seasonal parameters
}
\description{
Takes daily rain-forcing time series input and adds it to the parameter set
}
\examples{
n_days = 1000
t <- 1:n_days
rainfall_ts <- sin((t*2*pi)/365) + 1.1
params <- get_parameters() |>
          set_rainfall_manual(rainfall_ts) |>
          set_equilibrium(init_EIR = 5)
}
