% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{set_seasonality}
\alias{set_seasonality}
\title{Add seasonal patterns}
\usage{
set_seasonality(params, g0, g, h, floor = 0.001)
}
\arguments{
\item{params}{Other malariasimple parameters}

\item{g0}{Mean baseline coefficient}

\item{g}{Cosine coefficients}

\item{h}{Sine coefficients}

\item{floor}{Minimum value of seasonal forcing}
}
\value{
Updates the input parameter list to include seasonal parameters
}
\description{
Adds seasonal forcing to the model by approximating average seasonal rainfall as a Fourier series.
}
\examples{
# Define seasonality Fourier coefficients
g0 <- 0.28
g <- c(-0.3, -0.03, 0.17)
h <- c(-0.35, 0.33, -0.08)
params <- get_parameters() |>
          set_seasonality(g0=g0,
          g=g,
          h=h,
          floor = 0.005) |>
          set_equilibrium(init_EIR = 5)
}
