% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_spraying.R
\name{set_spraying}
\alias{set_spraying}
\title{Set IRS parameters}
\usage{
set_spraying(
  params,
  days,
  coverages,
  distribution_type = "random",
  ls_theta = 2.025,
  ls_gamma = -0.009,
  ks_theta = -2.222,
  ks_gamma = -1.232,
  ms_theta = -1.232,
  ms_gamma = -0.009
)
}
\arguments{
\item{params}{malariasimple parameters}

\item{days}{Days on which IRS is implemented}

\item{coverages}{Population coverage for each day}

\item{distribution_type}{"random" or "cohort" distribution of IRS}

\item{ls_theta}{Initial proportion of mosquitoes dying following entering a hut}

\item{ls_gamma}{Defines how ls changes over time}

\item{ks_theta}{Initial proportion of mosquitoes successfully feeding}

\item{ks_gamma}{Defines how ks changes over time}

\item{ms_theta}{Initial proportion of mosquitoes repelled}

\item{ms_gamma}{Defines how ms changes over time}
}
\description{
Set IRS parameters
}
