% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{get_custom_output}
\alias{get_custom_output}
\title{Get more specific outputs}
\usage{
get_custom_output(
  params,
  sim,
  output_variables,
  ages = NULL,
  biting_groups = NULL,
  int_groups = NULL
)
}
\arguments{
\item{params}{malariasimple input parameters}

\item{sim}{Output from malariasimple::run_simulation(params, full_output = TRUE)}

\item{output_variables}{Output variable of interest.}

\item{ages}{Age groups of interest (defined by the lower bracket)}

\item{biting_groups}{Biting groups of interest}

\item{int_groups}{Intervention groups of interest. 1 = No intervention, 2 = Bednets, 3 = SMC, 4 = Bednets and SMC}
}
\description{
Allows the user to filter to more specific human outputs. Aggregated values for any human compartment as well as
detected cases and clinical incidence may be requested for specific age, biting heterogeneity or intervention groups.
}
\examples{
  params <- get_parameters(
    n_days = 50,
    age_vector = c(0, 0.25, 0.5, 1, 3, 5, 10, 50) * 365) |>
    set_equilibrium(init_EIR = 10)
  sim <- run_simulation(params, full_output = TRUE)
  custom_output <- get_custom_output(
    params,
    sim,
    output_variables = c("n", "detect"),
    ages = c(0, 0.5) * 365)
}
