% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_peak_cc}
\alias{get_peak_cc}
\title{Timing of peak carrying capacity}
\usage{
get_peak_cc(season_params)
}
\arguments{
\item{season_params}{List of Fourier function parameters describing seasonality}
}
\description{
Returns the day-of-year on which carrying capacity is maximum
}
\examples{
season_params <- list(g0 = 0.28,
                          g = c(-0.30, -0.03, 0.17),
                          h = c(-0.35, 0.32, -0.08),
                          theta_c = 0.28)
peak_cc <- get_peak_cc(season_params)
}
