% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_peak_cc}
\alias{get_peak_cc}
\title{Timing of peak carrying capacity}
\usage{
get_peak_cc(g, g0, h)
}
\arguments{
\item{g}{Cosine coefficients}

\item{g0}{Mean baseline coefficient}

\item{h}{Sine coefficients}
}
\description{
Returns the day-of-year on which vector carrying capacity is maximum
}
\examples{
#Seasonality parameters
g0 = 0.28
g = c(-0.3, -0.03, 0.17)
h = c(-0.35, 0.32, -0.07)

peak_cc <- get_peak_cc(g0, g, h)
}
