% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{get_seasonal_forcing}
\alias{get_seasonal_forcing}
\title{Get daily seasonal forcing}
\usage{
get_seasonal_forcing(t, g0, g, h, floor = 0.001)
}
\arguments{
\item{t}{Day-of-year}

\item{g0}{Mean baseline coefficient}

\item{g}{Cosine coefficients}

\item{h}{Sine coefficients}

\item{floor}{Minimum permitted value of output}
}
\description{
Convert Fourier coefficients into a smooth vector of daily seasonal forcing. Used within set_seasonality function.
}
\examples{
# Define seasonality parameters
g0 = 0.28
g = c(-0.3, -0.03, 0.17)
h = c(-0.35, 0.33, -0.08)

daily_forcing <- get_seasonal_forcing(t = 1:365,
g0 = g0,
g = g,
h = h)
}
