% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameter_draws_df}
\alias{parameter_draws_df}
\title{Parameter draws (P. falciparum)}
\format{
## `parameter_draws_df`
A list of lists of length 1000, each level contains a list of drawn parameters (plasmodium falciparum)
}
\source{
<https://www.nature.com/articles/ncomms4136>
}
\usage{
parameter_draws_df
}
\description{
A list of 1000 draws from the joint posterior fit from
}
\keyword{datasets}
