% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Run deterministic malariasimple}
\usage{
run_simulation(gen, params, full_output = FALSE)
}
\arguments{
\item{params}{List of model parameters generated using get_parameters() and other helper functions}
}
\description{
Generates and runs the deterministic malariasimple model
}
\examples{
params <- get_parameters() |>
          set_equilibrium(init_EIR = 10)
gen <- gen_model()
simulation_output <- run_simulation(gen,params)
}
