% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality.R
\name{set_seasonality}
\alias{set_seasonality}
\title{Add seasonal patterns}
\usage{
set_seasonality(params, season_params, floor = 0.001)
}
\arguments{
\item{params}{Other malariasimple parameters}

\item{season_params}{List of Fourier parameters for manual seasonal input}

\item{floor}{Minimum value of seasonal forcing}

\item{admin_unit}{Name of admin 2 region}

\item{country}{Name of admin 2 country}
}
\value{
Updates the input parameter list to include seasonal parameters
}
\description{
Adds seasonal forcing to the model using a Fourier function of rainfall.
}
\examples{
season_params <- list(g0 = 0.28,
g = c(0.2, -0.07, -0.001),
h = c(0.2, -0.07, -0.1),
theta_c = 1)
params <- get_parameters() |>
          set_seasonality(season_params = season_params, floor = 0.005) |>
          set_equilibrium(init_EIR = 5)


}
